// components/event-calendar/types.ts
export type CalendarView = "month" | "week" | "day" | "agenda";

export interface CalendarEvent {
  id: string;
  title: string;
  description?: string;
  start: Date;
  end: Date;
  allDay?: boolean;
  color?: EventColor;
  location?: string;
}

export type EventColor =
  | "sky"
  | "amber"
  | "violet"
  | "rose"
  | "emerald"
  | "orange"
  | "red"
  | "blue"
  | "green"
  | "yellow"
  | "pink"
  | "indigo"
  | "cyan"
  | "teal"
  | "lime"
  | "fuchsia"
  | "slate";